move testapp tests to testapp and out of actstream
